Fix: On multiple reports, the data and dataset are overriden when have same definition
If you want to print an array of reports with the same definition (same report name), line 206 was referencing instead of copying the definition. So, you are modifying the data and dataset properties (line 222 and 223) for all and getting the last values of the last element of the loop (line 209). 

It is necessary to copy the definition and not get a reference.